fix: passed

+++ lib/percona_ar/query_builder.rb
@@ -11,8 +11,12 @@ class PerconaAr::QueryBuilder
   end
 
   def add(sql)
+    # DROP INDEX `index_studies_on_added_column` ON `studies`
     if sql =~ /^ALTER TABLE `([^`]*)` (.*)/i
       @tables[$1.to_s] << get_sql_for($2)
+    elsif sql =~ /DROP INDEX/i
+      drop_clause, table = sql.split(/ ON /i)
+      @tables[table] << get_sql_for(drop_clause)
     end
     self
   end
@@ -21,6 +25,7 @@ class PerconaAr::QueryBuilder
 
   def get_sql_for(cmd)
     return cmd unless cmd =~ /DROP/i && !(cmd =~ /COLUMN/i)
+    return cmd if cmd =~ /DROP INDEX/i
     cmd.gsub(/DROP/i, "DROP COLUMN")
   end
